fix chart style and refactoring
